+++ app/src/main/java/org/android/go/sopt/member/SignInActivity.kt
@@ -62,7 +62,6 @@ class SignInActivity : AppCompatActivity() {
                 signInResult.data.name ?: "",
                 signInResult.data.skill ?: ""
             )
-            //showToast(getString(R.string.login_success))
         }
     }
