+++ app/src/main/res/layout/activity_sign_up.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="org.android.go.sopt.member.MemberViewModel" />
+            type="org.android.go.sopt.member.SignUpViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
